+++ networkmonitor/src/main/java/alirezat775/lib/networkmonitor/OkHttpHelper.kt
@@ -1,4 +1,4 @@
-package alirezat775.networkmonitor
+package alirezat775.lib.networkmonitor
 
 import android.text.TextUtils
 import okhttp3.Request
@@ -82,12 +82,9 @@ object OkHttpHelper {
         ) {
             return true
         }
-        val contentLength = stringToLong(
-            response.header("Content-Length")
-        )
+        val contentLength = stringToLong(response.header("Content-Length"))
         return contentLength != -1L || "chunked".equals(
-            response.header("Transfer-Encoding"),
-            ignoreCase = true
+            response.header("Transfer-Encoding"), ignoreCase = true
         )
     }
